refactor(r1b): switch core callers to canonical mahamantra_research imports
- cli/kirtan_cli: research_chat -> vibe_core.mahamantra_research
- audit/gaps + audit/scale: project_introspection -> mahamantra_research
- audit/protocol_resurrection: acintya_mathematics -> mahamantra_research
- protocols/_gita_lens: gita imports -> mahamantra_research

Legacy compatibility path vibe_core.mahamantra.research remains via shim.
