+++ vibe_core/mahamantra/audit/scale.py
@@ -28,7 +28,7 @@ def get(root: Path = None) -> Dict[str, Any]:
     Returns:
         Dict with: total_files, total_lines, coverage_percent, broken_lineage
     """
-    from vibe_core.mahamantra.research.project_introspection import measure_scale
+    from vibe_core.mahamantra_research.project_introspection import measure_scale
     return measure_scale(root or Path.cwd())
